+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **You can now click and drag** with the "Wink to click" and "Open mouth to click" modes. You may find it harder to perform simple clicks due to the mouse button being held down, but this will probably be addressed before the next release with an option.
+- **Motion Threshold** slider, similar to the setting in [eViacam](https://eviacam.crea-si.com/). This helps keep the mouse still when you stop moving your head. This is especially useful for single and double clicking in clicking modes that allow dragging, but is also useful for dwell clicking. (You might want to disable it if you're going to be drawing on a canvas.)
 - Blink detection includes a visualization in the camera view. It may make it look like you're wearing glasses. 😎
 - Open mouth detection includes a visualization in the camera view. This is drawn as two lines for now.
 - Blink detection and open mouth detection also include a visualization near the mouse cursor, so you can be confident when it's clicking, even if what you're clicking on doesn't provide any visual feedback.